backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
psycopg[binary]==3.1.19
redis==5.0.4
django-environ==0.11.2
